requirements.txt
python-dotenv>=1.0.1
PyYAML>=6.0.1
pydantic>=2.7.0
openai>=1.51.0
pypdf>=5.0.0
python-docx>=1.1.2
beautifulsoup4>=4.12.3
markdownify>=0.13.1
